Initial code commit
Added the header and implementation for quick sort, heap sort, merge sort, insertion, and bubble sort. Also included the implementation and header for a simple timer class capable of ranking the sorting algorithm run times. A main.cpp is included to test out the functions via command line. A makefile to compile the code via g++ is also provided.